hide internals
